Refactored CSV Formatter classes
