+++ Plugins/FormatCsv/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/format/csv/tools/NavigationPropertyFollower.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2020 Fraunhofer Institut IOSB, Fraunhoferstr. 1, D 76131
+ * Karlsruhe, Germany.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package de.fraunhofer.iosb.ilt.frostserver.plugin.format.csv.tools;
+
+import de.fraunhofer.iosb.ilt.frostserver.model.core.Entity;
+
+/**
+ * Interface for a class that follows a navigation link and returns the linked
+ * Entity.
+ *
+ * @author scf
+ */
+public interface NavigationPropertyFollower {
+
+    /**
+     * Follow the registered navigation link, on the given source Entity, and
+     * return the linked Target Entity.
+     *
+     * @param source The source Entity.
+     * @return The linked target Entity.
+     */
+    public Entity<?> fetch(Entity<?> source);
+
+}
